requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
openai==2.16.0
mcp==1.26.0
colorlog==6.9.0
librosa==0.11.0
transnetv2_pytorch==1.0.5
ffmpeg-python==0.2.0
aiofiles==23.2.1
moviepy==2.2.1
av==16.1.0
sentence-transformers==5.2.2
tomli
faiss-cpu==1.13.2
emoji==2.15.0
Pillow
numpy
